requirements.txt
asgiref==3.7.2
dj-static==0.0.6
Django==5.0.1
django-cleanup==8.0.0
django-js-asset==2.2.0
django-tinymce==3.6.1
pillow==10.2.0
sqlparse==0.4.4
static3==0.7.0
typing_extensions==4.9.0
tzdata==2023.4
